Indent line comments correctly when they precede a continuation line.
If a line comment occurs after a continuation break, then its indentation
depends on what comes on the line following it:

* If the following line is an indented continuation line, then the comment
  should also be indented as a continuation. For example,

  ```swift
  return true
    // comment
    && false
  ```

* Otherwise, if the following line is not a continuation (such as a new
  statement, a closing delimiter, or the end of the file), then the comment
  should not be indented as a continuation:

  ```swift
  {
    return true
    // comment
  }
  ```

This is achieved by adding a reset break after the last item in a code
block (rather than just _between_ them) and by not resetting the
continuation state unconditionally when printing a comment.

Fixes SR-11107.