Ensure line breaks between type members wrapped in an `#if` block.
These were getting collapsed because we were adding the breaks inside
the visitor for `MemberDeclBlock`, but when there is an intervening
`IfConfigClause`, the inner members are only wrapped in a
`MemberDeclList`, not another `MemberDeclBlock`. Updating the token
stream to insert the breaks in the list visitor ensures that this
happens in all cases.

Fixes SR-11116.