Preserve order of case elements in OneCasePerLine.
The old implementation of the rule would remove associated/raw value
elements from the case declaration that they were in and move them to
their own case declaration below it. This is incorrect for cases with
numeric raw values, because it changed their relative order.

The new implementation scans the cases/elements linearly and emits
new declarations when needed during the scan, keeping the original
order.

This also cleans up the way comments on cases are handled and ensures
that attributes on cases work correctly as well.

Fixes SR-11109.