Make lint rules extend SyntaxVisitorBase.
In an older version of SwiftSyntax, SyntaxVisitor was a class and thus
the lint rules were as well. When SwiftSyntax changed SyntaxVisitor to be
a protocol with mutating methods, we lost the ability to treat the rules
polymorphically in certain situations due to Swift language limitations.

Now that SwiftSyntax generates a class version of SyntaxVisitor, we can
bring that old design back.