Fix rule enabling/disabling, both in config and in source comments.
RuleMask previously only kept track of rule disablement, assuming that
any range where a rule was not disabled should have it enabled. This
isn't quite correct; it's actually a tri-state of enabled, disabled, and
whatever-is-in-the-configuration.

With these changes, rule enablement/disablement is now handled in the
lint pipeline (format pipeline coming later), not in individual rules.
(This will actually be refined later, because some individual rules do
need finer control over that check.)

Fixes SR-11114.