Improve method chaining, especially involving trailing closures.
There are still some edge cases that can look odd when discretionary
newlines are involved and when wrapping also occurs around them, but
they can be repaired manually for now (and the formatter will preserve
that correction as long as it's also correct according to the rules).
Improving that will require a bit more work.

Fixes SR-11169.