Migrate leading trivia when adding an access-level modifier.
The AddModifierRewriter was previously only removing the leading trivia of the declaration's first token. The assumption that the new modifier must be the declaration's first token is incorrect. Instead, the AddModifierRewriter now transfers trivia from the token immediately following the new modifier without making any assumptions about where the modifier will be placed in the declaration. This transfer of trivia only occurs for the first modifier in a declaration. The code that adds modifiers to an existing list of modifiers already correctly moves trivia when adding modifiers.

This fixes a bug where leadingTrivia that belongs on attributes (e.g. `@objc`) was moved to the access-level modifier. Moving the leadTrivia trivia resulted in breaking up the attributes and modifiers in a declaration.