Allow line breaks between Obj-C selector pieces in attributes.
Previously, Objective-C selector pieces were forced onto 1 line regardless of their length. This is problematic when the selector is verbose or has many arguments, since it will overflow the line length. It's still possible for an Objective-C selector piece whose name is longer than the configured line length to overflow the line length, since it's invalid to break inside of the name of a selector piece.

These breaks are specifically size 0 because Objective-C selector pieces aren't separated by spaces.