Prevent extra indentation for trailing closures in specific function calls.
Certain kinds of multi-line function arguments were causing trailing closures to have an extra level of indentation when the closure was across multiple lines. This was caused by treating the last line of the function call, where the closure started, as a "continuation". The behavior to treat that line as a continuation allows `reset` breaks to work correctly, but it also means `open` breaks receive an extra level of indentation. See commit 40ab473 for details and background.

To fix this, I've changed the pretty printer to only treat these lines are continuations if the original open contributed to indentation (`didIndent == true`). Intuitively, this means a `close` break only creates a continuation if it's closing a "scope" that was indented.