Indent continuation lines in conditions for if & guard statements.
Previously, the conditions in if & guard statements were indented solely using continue breaks. Per design, continue breaks don't stack to create increasing line indentation. In order to increase the indentation by more than 1 indent, the pretty printer requires open & close break tokens.

This change wraps specific condition elements in open & close break tokens so that each condition has an indentation stack, and continuations inside of the condition element correctly indent relative to the start of the condition. There is some special handling for if-statements, due to a reasonable desire to avoid line breaks following the "if" token.

Note: It looks like there are some new tests that weren't included in the XCTestManifests, specifically the "SemiColonTypeTests". These tests were picked up by the generating code when I updated the XCTestManifests for my new tests. I've left it as-is since this needs to be fixed anyway. I'm happy to pull it into a separate change if that's preferred.