Improve handling of comments when collapsing fallthrough only switch cases.
There are two improvements:
- Avoid collapsing fallthrough statements that have comments.
- Move displaced trivia from collapsed fallthrough statements to avoid deleting comments.

Without moving the trivia, inline comments before the collapsed cases were being deleted. That is a destructive change that shouldn't happen.

Even though fallthrough only cases with comments could be collapsed, the presence of a comment in addition to the fallthrough indicates there's some value to describing the fallthrough. Migrating the comment would be difficult to get right and deleting comments is destructive. It's better to keep the fallthrough and its comments in this case.