Allow for no blank line between members when there's a preceding comment + new line.
When the members are preceded by a comment + new line, the comment is probably not directly attached to the first member. Instead, it's likely to be a navigational or organizational construct, such as a "MARK" comment. This change allows members following such a comment to not have a blank line between them.