Detect and delete semicolons nested in code blocks.
The DoNotUseSemicolons rule previously only removed semicolons that were on the top level code block items inside of a code block item list. It's possible for additional code block item lists to be nested inside of a code block item; see the new test for an example.

The rule now recursively walks through all children of each code block item to find and remove semicolons.

The leading trivia replacement to add a newline was stripping all newlines from the existing trivia, and then adding 1 newline. This is a problem when the leading trivia already contains 2 or more leading newlines, for example to create a blank line between 2 code block items. The previous behavior would remove the intended blank line. Instead, now a single newline is appended to the leading trivia if and only if it doesn't already contain at least 1 leading newline.