Fix extra indentation on closure variables.
Extra indentation was applied to closures because the following sequence of tokens occurs when the closure has a signature:`break(.continue), "{", newlines, break(.open)`. When the `break(.open)` is encountered, the pretty printer pushes 2 indents onto the `indentationStack`. The newline between the breaks causes the pretty printer to incorrectly record that the closure's first line break occurs in a continuation line.

Other structures with braces, including closures without a signature, avoid this problem by inserting the `break(.open)` after their left brace instead of inside of the structure. I have moved the break so that it always precedes any newlines inside of the closure's braces.