Expression grouping improvements to relocate breaks using a new "open-continue-break".
Some types of expressions were breaking at suboptimal locations in the source as a result of a lack of groups. I experimented with grouping around every expression in a SequenceExpr, but there are so many kinds of expressions used in so many contexts that it's difficult to determine appropriate groups in general. A second complicating factor is that the precedence of operators cannot be determined using the parsed syntax. Without knowing operator precedence, it's difficult to correctly group operators in general.

This change uses the following assumptions to make some progress on expression grouping:
1. Logical AND and OR operators (&&, ||) are generally good candidates for breaks and their operands are generally best grouped together.
2. Choices in ternary expressions are generally best grouped together, so that the semantic tokens in the ternary ("?", ":") are easier to find.
3. Member access expressions and function call expressions generally look best grouped together, allowing the breaks to occur at other expressions.

I have added a few new test cases that exercise the revised formatting behaviors, and updated the few existing test cases that were failing to match the new formatting.