Add a SequenceExpr folding algorithm.
This change adapts an algorithm used by the compiler to fold
sequence expressions into expression trees based on the precedence
and associativity of operators.

We will use this in TokenStreamCreator to transform SequenceExprs
into shapes that let us more easily insert groups and breaks that
allow wrapping to occur in a manner consistent with precedence and
associativity, which matches users' expectations.