Fix formatting when respectsExistingLineBreaks is false.
At some point, we accidentally broke formatting in this mode (when
ignoring user-discretionary line breaks) and it turns out we didn't
have any testing with that mode applied to catch this. This change
fixes it, strengthens the old support to handle semicolon-delimited
statements/declarations, and adds some regression tests to make sure
we don't break again.