Remove false warnings from synthesized initializer rule.
The synthesized initializer rule didn't consider two rules of the memberwise initializer for structs:
- The synthesized initializer is only available when there are no initializers.
- The synthesized initializer provides default arguments that exactly match the default values of corresponding properties.

The rule verifies that the struct only contains initializers that can be replaced by synthesized memberwise initializer(s). Otherwise, there is no warning because the synthesized initializer isn't created for structs that have any custom initializers and the custom initializers cannot be removed.

The rule verifies that any default arguments on the parameters match default values on corresponding properties. There was support for verifying that a matching default argument exists on a parameter whose property has a default value, but all other default arguments were ignored.

Finally, I revised the test suite for this rule to independently verify the various conditions that do and do not trigger the diagnostic.