Gracefully transition from function call -> tuple expr.
Recent changes in the syntax tree folded function call arguments into
tuple expressions. This change introduces some conditional compilation
directives that let us build under old and new toolchains while we
migrate some infrastructure.

This was tested with the 09-26 dev snapshot and the 10-21 snapshot
(which explains the need for the GenericRequirement condition, as it
was reverted during that time and only put back after the 10-21
snapshot was released).