Don't group around function calls that occur after an assignment operator.
Grouping around these function calls results in suboptimal formatting for expressions that simply assign the result of a function call. There was also inconsistent between assignments and declarations (e.g. using let or var), because there may be SequenceExpr when using let/var decl.

I opted to also exclude the grouping when the function call is the first item in an expression list, because the indentation of following items isn't confusing whereas the indentation would be confusing when the function call follows other operands.