Wrap after assignment operators, not before.
This required some changes to the pretty-printer around newlines
that are adjacent to breaks that I'd like to revisit at some point
in the future; there are a lot of subtleties in situations involving
breaks immediately before a newline vs. immediately after a newline
that I think could be simplified and cleaned up.