Add a test to verify `Configuration` initialization paths.
When initializing a `Configuration` with the default initializer,
it should have the same values as one decoded from an empty JSON
dictionary. Testing this lets us verify that the default values
assigned for missing keys match the default values assigned when
creating a new configuration.

This change also makes `Configuration` into a value type instead
of a class.