Use a closure instead of a keypath in a call to `map`.
I accidentally snuck in a usage of the keypath-to-function-literal
conversion (SE-0249) since master is developed against a trunk
development snapshot, but this feature isn't part of any released
Swift. It's not so critical that we need it, so it's better to just
use a closure for now so that this change can be cherry-picked back
into the 5.1 branch.