Allow the closing paren for function decls in protocols on the same line as the last parameter.
Breaking before this paren doesn't aid in the readability in the same way it does when the decl isn't part of a protocol, because there's no function body nor curly braces after the paren in this case. This break existed to aid readability by creating a boundary between the function's signature and its body. When the decl is part of a protocol, there is no body and the boundary isn't necessary.

In order to make this change, I had to hoist the grouping and breaking around parens in a function signature from the ParameterClauseSyntax visitor into parent visitor methods. The ParameterClauseSyntax is nested below the node which can determine if the decl has a body. Instead of traversing parent pointers, I hoisted the implementation into the parent visitors.