Put freestanding `init` decls in a test into a type.
Even though these aren't semantically valid Swift, these parse correctly
with the 9/26 snapshot we're currently on. At a more recent commit we
noticed that these `init`s seem to be hitting a code path that causes an
assertion in the parser:

    assertion failed in swift::Type swift::ValueDecl::getInterfaceType() const:
    ctx.getLegacyGlobalTypeChecker() && "The type checker must be installed to
    make semantic queries!"

We should try to pinpoint this crash, but for now I'm patching this so
that the test is more "correct" and it unblocks some of our internal work.