Support tab based indentation in the whitespace linter.
The linter was crashing whenever it encountered a tab because it only expected spaces to be used for whitespace. The linter now supports a wider range of whitespace characters and allows using tabs for indentation. This includes some revisions to the error messages to add tabs are as a suggestion for adding/removing indentation. There is also a new error message used when tabs (or really any non-space) characters are used for spacing (i.e. the whitespace that isn't indentation).

SR-11211