Use continuation opens instead of block opens around conditions in if and guard stmts.
The block type of open creates 2x indents when the open break occurs on a line that is a continuation, because the relevant active-open-break can contribute a block indent and a continuation indent. The conditions in if and guard stmts aren't code blocks. These open breaks were added to support stacking a single continuation inside of each condition. The continuation type of open was added later and better models the behavior desired in this case.