Add open breaks around type annotations in optional condition exprs.
This matches the breaks and grouping used by pattern binding syntax nodes (i.e. let and var expressions) so that the formatting of a let in a guard/if stmt will have the same formatting as an identical let/var expression outside of a guard/if stmt.

This fixes an issue where there was no space between the ":" and the type annotation when the let in an optional condition included a type annotation.