Improve wrapping of parenthesized expressions.
Since paren exprs are just represented as one-element tuples, they
were getting the same kind of wrapping as tuples, with breaks inside
the parens. This doesn't look great, because these are different kinds
of expressions. Now, we keep the parens inseparable from the tokens
that they are adjacent to, which lets them wrap more traditionally,
and we increase the continuation indentation inside them.

This required some changes to the way we stack indentation for
variable initialization and assignment to make multiple parenthesized
expressions line up nicely. For example, whereas we used to do this:

```swift
x =
  foo
    + bar
    + baz
```

Now we do

```swift
x =
  foo
  + bar
  + baz
```

...so that if those terms are paren exprs and they wrap inside, they
still shift in and out nicely and line up well:

```swift
x =
  (foo
    + foo2)
  + (bar
    + bar2)
  + (baz
    + baz2)
```

We find the loss of the original extra level of indentation to be
acceptable because the subsequent lines still start with a spaced
operator or some other token (like a prefix dot) that indicates
that the line is unambiguously a continuation line.