Reverse `.break(.close)` and `.close` for array and dict literals.
The previous order created an issue where the `.break(.close)` was outside of the group. The result was that it was possible just for the closing delimiter for an array/dict could break off from the rest of the literal. The new order of tokens places the break inside of the group, so the `.break(.open)` before the literal includes the length of the array/dict elements and the closing delimiter.