Diagnostics for trailing comma changes.
As part of the migration from a phase 1 rule for trailing commas to using the PrettyPrinter, we still need to emit diagnostics when the source's use of trailing commas is incorrect. These diagnostics are similar to those from the original phase 1 rule, except they don't specify the type of collection (instead saying "collection literal"). I think context of the line and column number + the general term "collection" are sufficient for the user to understand the intent.

I added support to PrettyPrintTestCase for collecting diagnostics so that they can be tested.

Finally, the WhitespaceLinter is only able to handle changes in whitespace. By adding/removing trailing commas, the whitespace linter would fail. Instead of checking the whitespace linter to support skipping comma changes, I've created a "mode" in the PrettyPrinter to print the text of the source exactly as it was while only changing whitespace. If we ever introduce other types of text edits in the PrettyPrinter, this should work for that too.