No trailing commas in array/dict initializers.
swift-syntax doesn't differentiate the array/dict expr syntax used for initializing an array/dict (where the elements are type identifiers) from an array/dict expr syntax of a collection literal (where the elements are the contents of the collection). There isn't a straightforward way to determine which is which when visiting the array or dict, since it's based on the parent being a function call.

A simple solution is to never add a trailing comma to array/dict when there is only 1 element. This impacts multiline array/dict literals with only 1 element, but that's probably uncommon and it's reasonable to skip the trailing comma for only 1 element.