Diagnose using line & col of formatted text in PrettyPrinter.
The PrettyPrinter was using `AbsolutePosition`s from the TokenStreamCreator when creating diagnostics. Those are based on a utf8 offset, which is invalidated when the PrettyPrinter adds/removes content from the text while printing it. Instead, the diagnostics from the PrettyPrinter use a line and column computed based on its output buffer.

I removed all instances of storing `AbsolutePosition` from TokenStreamCreator, since using `AbsolutePosition` in PrettyPrinter is inherently flawed.