Preserve garbage text trivia when pretty printing.
Garbage text typically refers to unrecognized characters in the input,
which we haven't typically cared about. However, there are also some
uses of garbage text that don't represent "garbage" per se, but content
that the lexer/parser doesn't care about but which we should still
preserve during formatting:

* a hashbang line at the beginning of the file
* a Unicode byte-order marker at the beginning of the file
* source control conflict markers

When we see these (or any garbage), we treat them as verbatim content
so that we can handle it as nondestructively as possible.

This change also makes some small fixes to verbatim handling, and
removes a convenience constructor for verbatim tokens so that we force
the caller to specify the indentation behavior instead of relying on an
arbitrary default.