Remove BlankLineBetweenMembers rule because it creates non-idempotent behavior.
The rule is unfortunately based on the trivia before the pretty printing pass, which means it decides single-line-ness based on the input which may be incorrect. The single-line-ness must be based on the source *after* pretty printing, so it cannot be accomplished in a phase 1 rule.

Even linting in a phase 1 rule is too inaccurate, because it may flag false positives and false negatives creating useless noise.

In the future, the blank line between members requirement can be better implemented using breaks in the pretty printer.