Swap order of .close and .break around tuple exprs.
The close group and close break were in the wrong order, causing the right paren's length to not be included in the length of the open break before first tuple element. This allowed the open break between the left paren and first element to NOT fire, but later the close break before the right paren DOES FIRE due to line length.