Transform nested structs in UseEnumForNamespacing.
The rule was not visiting any structs nested in other structs, because it wasn't calling super to visit the node's children.

I tried to combine `super.visit` with the member enumeration in `membersToKeepIfUsedAsNamespace(_:)` but that method intentionally returns nil to signal that the decl shouldn't be rewritten which would discard the changes to the nested nodes. Instead, the rule now visits the members and uses the rewritten member list even if it's not converted to an enum to keep any changes to nested structs.