Delete unsafe UseEnumsForNamespacing rule
While using enums for namespaces is a superior coding practice, this rule lacks sufficient context to safely perform this transformation. It's possible to use extensions in the same file or a different file to add a conformance to a struct which makes it ineligible to become an enum. Unfortunately, the SyntaxVisitor is only able to look at nodes in the AST for a single file and that is too narrow of a scope to verify that the struct can safely be an enum.

Even as a lint rule, this would raise false positives and those false positives cannot be removed. It's best not to convert to a lint rule for that reason.