Use a location specific diagnostic message when the input is invalid.
Previously, the diagnostic didn't include a location (i.e. filename, line number, and column number) so the diagnostic message included just the filename. Now the diagnostic is added at a location so that it's more consistent with other diagnostics and includes additional useful information in the same format as other diagnostics.

The diagnostic for an invalid file now looks like:
/path/to/file.swift:line:column: error: file contains invalid or unrecognized Swift syntax.