Add a group around member types used in return clauses.
Without the group, the pretty printer would add a newline at the first break in the return clause where the column limit would overflow. This could fall inside of a member type identifier, before the dot. That results in odd formatting where a func decl has its return clause partially on the first line and partially on the second line (which is only allowed for tuples because of the opening paren special case).

With the group, the func decl breaks before the -> when the type would overflow the line.