Support an ignore directive to ignore an entire file.
This is intentionally a different ignore directive (`swift-format-ignore-file` instead of `swift-format-ignore`) so that there won't be any confusion if someone wants to ignore formatting on the first node in a file (e.g. ignore the first import) but wants to format the rest of the file. If ignore file used the same comment as ignore a node, users would need to artificially structure their code in a way that avoids putting an ignore on the first node in the file when they want part of the file to be formatted.