Don't propagate the diagnostic engine when formatting.
This supresses printing diagnostic messages when using the format mode. When formatting, the issues in the source are either going to be fixed or they can't be fixed automatically. In either case, the formatter should be quiet about it.

Linting continues to use the `DiagnosticEngine` so that lint findings are displayed as before.