Restore some discretionary newlines from PR #143
Several breaks were changed to disallow discretionary newlines in #143. In practice, some of these breaks actually require a more nuanced approach because grouping around the content after the `:` is frequently subjective. Depending on the length of the label/key before the `:` and the actual expr or type, it varies whether a group should be around the content after the `:`. The formatter's current model doesn't support modifying the tokens to add/remove group tokens once the token stream is created.

In a future patch, it would be nice to have a way to switch between grouping and not grouping the content after the `:` based on the preceding key or label.

Reverted `ignoresDiscretionary` breaks include after the colon in...
- a dictionary type
- a dictionary element
- a function call argument
- a function decl parameter
- a generic parameter
- a tuple type
- a tuple expr element