Add breaks around selector expressions.
I've found 2 uses for `DeclNameArgumentSyntax`: in an Objective-C selector (`#selector`) and when a function is referenced.

I decided not to use any of the various arrangeAsFunction... methods even though this is similar to a function call. This syntax is much simpler than a function call because it only contains the labels, so there's no need for any of the logic for grouping around function call arguments or types.