Migrate modes to subcommands.
This change refactors the command line interface's
modes from using a flag (`-m/--mode`) to using
swift-argument-parser subcommands instead. This
provides a much better separation of responsbilities
because subcommands only define arguments that they
support, rather than having a single command that
supports the union of all possible arguments.

The old mode flag is still supported for backward
compability (see `LegacyMainCommand`) in order to
give clients who may be using it in automation an
opportunity to migrate away from it, but it will
be removed in the future.